Add Enemy Hp System

+++ Assets/Scripts/EnemyAttackSO.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public abstract class EnemyAttackSO : ScriptableObject
 {
     [Header("Attack Settings")]
-    public float enemyDamage;
+    public int enemyDamage;
     public float enemyAttackRange;
     public float enemyAttackDelay;
     public float enemyAttackDuration;
